Creating documents for HTML diagnostics production. This stuff may move into a specific project repository at some point
These HTML diagnostics are just for referential integrity and are completely undocumented currently. Plans are to create another version that works within oXygen or if you have Saxon EE or another Saxon that allows you to use the file: functions (currently not available with HE). Some of these files are to be deleted in a later commit, once the build process is finished.